+++ NEWS.md
@@ -15,6 +15,9 @@ to integer
 be assigned in case it is missing and an error will continue to be generated
 when dosing into an invalid compartment (0 is always an invalid index
 for dosing compartment)
+- An error will be generated if missing values (`NA`) are detected in an 
+input data set with the following names: `ID`, `time/TIME`, `rate/RATE` ,
+or any column that shares a name with an item in the parameter list.
 
 # mrgsolve 0.8.12
 - Minor changes to namespace for CRAN